+++ packages/desktop/src/renderer/components/SearchModal/SearchModal.test.tsx
@@ -76,7 +76,7 @@ describe('Search Modal', () => {
             aria-hidden="true"
           />
           <div
-            class="MuiModal-root css-1on48p8-MuiModal-root"
+            class="MuiModal-root css-1l68gny-MuiModal-root"
             data-testid="searchChannelModal"
             role="presentation"
           >
@@ -116,6 +116,7 @@ describe('Search Modal', () => {
                     >
                       <button
                         class="MuiButtonBase-root MuiIconButton-root IconButtonroot MuiIconButton-sizeMedium css-1hpikoh-MuiButtonBase-root-MuiIconButton-root"
+                        data-testid="ModalClose"
                         tabindex="0"
                         type="button"
                       >