+++ packages/desktop/src/renderer/components/ui/ErrorModal/ErrorModal.test.tsx
@@ -23,7 +23,7 @@ describe('ErrorModal', () => {
           aria-hidden="true"
         />
         <div
-          class="MuiModal-root css-1on48p8-MuiModal-root"
+          class="MuiModal-root css-1l68gny-MuiModal-root"
           role="presentation"
         >
           <div
@@ -64,6 +64,7 @@ describe('ErrorModal', () => {
                   >
                     <button
                       class="MuiButtonBase-root MuiIconButton-root IconButtonroot MuiIconButton-sizeMedium css-1hpikoh-MuiButtonBase-root-MuiIconButton-root"
+                      data-testid="ModalClose"
                       tabindex="0"
                       type="button"
                     >