+++ packages/mobile/src/components/ConnectionProcess/ConnectionProcess.test.tsx
@@ -39,7 +39,7 @@ describe('ConnectionProcessComponent', () => {
             collapsable={false}
             source={
               {
-                "testUri": "../../../assets/icons/join-community.png",
+                "testUri": "../../../src/assets/icons/png/join-community.png",
               }
             }
             style={
@@ -176,7 +176,7 @@ describe('ConnectionProcessComponent', () => {
             }
             verticalTextAlign="center"
           >
-            Quiet stores data on your community’s devices (not Big Tech’s servers!) using the battle-tested privacy tool Tor to protect your information. Tor is fast once connected, but it can be slow at first, and leaving this screen
+            Quiet stores data on your community’s devices using the battle-tested privacy tool Tor to protect your information. Tor is fast once connected, but it can be slow at first, and leaving this screen
              will stop the process of joining.
           </Text>
           <Text